Remove asv dev command
