+++ asv/commands/dev.py
@@ -21,6 +21,12 @@ class Dev(Run):
 
     @classmethod
     def run(cls, conf, **kwargs):
-        if not kwargs.get("env_spec"):
-            kwargs["env_spec"] = ["existing:same"]
-        return super(cls, Dev).run(conf=conf, **kwargs)
+        raise RuntimeError(
+            "`asv dev` has been removed. It was a shortcut of `asv run`, which "
+            "you can use instead. It was removed because it caused confusion, "
+            "in particular after changing what it was running in asv 0.5.\n\n"
+            "You can be interested in the next `asv run` arguments:\n"
+            " - `--python=same`: to use the current environment and not create one\n"
+            " - `--quick`: to run benchmarks just once\n"
+            " - `--dry-run`: to not save the results\n"
+        )
